+++ CHANGELOG.markdown
@@ -32,6 +32,8 @@
 - Support for 2-4 -element float sliders
 - `ImVec4::zero()`
 - `Into` array and tuple conversions for ImVec2 and ImVec4
+- gfx 0.14 support in imgui-sys
+- gfx 0.14 renderer implementation
 
 ### Changed